perbaikan file migrate lagi

+++ database/migrations/2026_07_06_020700_create_it_wh_todos_table.php
@@ -11,17 +11,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('it_wh_todos', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('assigner_id')->constrained('users')->onDelete('cascade');
-            $table->date('date');
-            $table->string('task_name');
-            $table->date('deadline');
-            $table->string('status')->default('To Do'); // 'To Do', 'In Progress', 'Done'
-            $table->text('notes')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('it_wh_todos')) {
+            Schema::create('it_wh_todos', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('assigner_id')->constrained('users')->onDelete('cascade');
+                $table->date('date');
+                $table->string('task_name');
+                $table->date('deadline');
+                $table->string('status')->default('To Do'); // 'To Do', 'In Progress', 'Done'
+                $table->text('notes')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
